Add code with fixed service classes
