+++ src/main/java/ua/lviv/iot/coursework/controllers/PanelController.java
@@ -3,7 +3,7 @@ package ua.lviv.iot.coursework.controllers;
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.web.bind.annotation.*;
 import ua.lviv.iot.coursework.models.SolarPanel;
-import ua.lviv.iot.coursework.logic.panel.PanelService;
+import ua.lviv.iot.coursework.logic.panel.impl.PanelServiceImpl;
 
 import javax.websocket.server.PathParam;
 import java.util.List;
@@ -11,41 +11,41 @@ import java.util.List;
 @RestController
 @RequestMapping("/solarpanel")
 
-public class PanelController extends PanelService {
+public class PanelController extends PanelServiceImpl {
 
     @Qualifier("PanelService")
-    private final PanelService panelService = new PanelService();
+    private final PanelServiceImpl panelServiceImpl = new PanelServiceImpl();
 
     @PostMapping
     @Override
     public void create(@RequestBody SolarPanel solarPanel) {
 
-        panelService.create(solarPanel);
+        panelServiceImpl.create(solarPanel);
     }
 
     @GetMapping
     @Override
     public List<SolarPanel> readALL() {
 
-        return panelService.readALL();
+        return panelServiceImpl.readALL();
     }
 
     @GetMapping("/{id}")
     @Override
     public SolarPanel read(@PathParam("id") @PathVariable("id") int id) {
 
-        return panelService.read(id);
+        return panelServiceImpl.read(id);
     }
     @PutMapping("/{id}")
     @Override
     public boolean update(@PathVariable int id, @RequestBody SolarPanel solarPanel) {
 
-        return panelService.update(id, solarPanel);
+        return panelServiceImpl.update(id, solarPanel);
     }
 
     @DeleteMapping("/{id}")
     @Override
     public boolean delete(@PathVariable int id) {
-        return panelService.delete(id);
+        return panelServiceImpl.delete(id);
     }
 }
