+++ src/main/java/ua/lviv/iot/coursework/logic/sensor/SensorService.java
@@ -1,66 +1,14 @@
 package ua.lviv.iot.coursework.logic.sensor;
 
-import org.springframework.stereotype.Service;
-import ua.lviv.iot.coursework.csvmanagers.SensorCSVManager;
-import ua.lviv.iot.coursework.models.PanelOwner;
 import ua.lviv.iot.coursework.models.Sensor;
-import ua.lviv.iot.coursework.logic.sensor.impl.SensorServiceImpl;
 
-import java.util.LinkedList;
 import java.util.List;
 
-@Service("SensorService")
-public class SensorService implements SensorServiceImpl {
+public interface SensorService {
 
-    SensorCSVManager manager = new SensorCSVManager();
-
-    @Override
-    public void create(Sensor sensor) {
-
-        var checkIfEmpty = new LinkedList<Sensor>();
-        checkIfEmpty.add(manager.readHash(1));
-        if (checkIfEmpty.isEmpty()) {
-            manager.addDataToHashFromCSVFile();
-        }
-        manager.putToHash(sensor);
-    }
-
-    @Override
-    public List<Sensor> readALL() {
-
-        var checkIfEmpty = new LinkedList<Sensor>();
-        checkIfEmpty.add(manager.readHash(1));
-        if (checkIfEmpty.isEmpty()) {
-            manager.addDataToHashFromCSVFile();
-        }
-        return new LinkedList<Sensor>(manager.getAllHash());
-    }
-
-    @Override
-    public Sensor read(int id) {
-
-        var checkIfEmpty = new LinkedList<Sensor>();
-        checkIfEmpty.add(manager.readHash(1));
-        if (checkIfEmpty.isEmpty()) {
-            manager.addDataToHashFromCSVFile();
-        }
-        return manager.readHash(id);
-    }
-
-    @Override
-    public boolean update(int id, Sensor sensor) {
-
-        var checkIfEmpty = new LinkedList<Sensor>();
-        checkIfEmpty.add(manager.readHash(1));
-        if (checkIfEmpty.isEmpty()) {
-            manager.addDataToHashFromCSVFile();
-        }
-        return manager.updateHash(id, sensor);
-    }
-
-    @Override
-    public boolean delete(int id) {
-
-        return manager.removeFromHash(id);
-    }
+    public void create(Sensor sensor);
+    public List<Sensor> readALL();
+    public Sensor read(int id);
+    public boolean update(int id, Sensor sensor);
+    public boolean delete(int id);
 }
